refactor(tagger): 抽 OnnxTaggerBase 基类，CLTagger 自动获得 PP10 线程池
WD14 + CLTagger 各自实现的 CUDA fallback、batch 决议、ThreadPoolExecutor
预处理、tag() batch 循环骨架、safe_dir_name / silenced_fd_stderr 都统一到
studio/services/onnx_tagger_base.py:OnnxTaggerBase。子类只填 prepare /
_preprocess / _postprocess_one / _get_batch_size_cfg 四个钩子。

效果：
- WD14 减约 190 行；CLTagger 减约 110 行。
- CLTagger 自动拿到 PP10 的预处理线程池（之前漏写，batch>1 时单核 PIL 是
  瓶颈）。
- CLTagger CUDA 成功路径会调 record_cuda_load_error(None) 清旧错状态
  （之前漏写，会把 WD14 的旧 CUDA 红字一直挂在 Settings UI 上）。
- safe_dir_name / silenced_fd_stderr 改成公开 API，model_downloader 不再
  跨模块 import 私有符号 + inline replace。

附带把 PR #14 cltagger_tagger.py 几处未独立成 commit 的小修一起带上：
- _compute_base() 替代 model_path.parents[len-1] 反推（写法脆，模型路径
  末尾带斜杠就算错）。
- 下载后 model / mapping 缺失分别列出，错误信息指向 Settings 下载日志。
- 移除 np.nan_to_num：_sigmoid 内 clip(±30) 已经吞 ±inf；NaN 经 sigmoid
  仍是 NaN，threshold 比较自然 False，等同丢弃。
